Fix overflow of reported size above ~2GB
The return type of `get_size` was only guaranteed to be 32 bits (and
not signed) which overflows at around 2GB. By replacing it with a
`long long` which is guaranteed to be 64 bits, we fix the overflow.

+++ src/gzsize.c
@@ -6,7 +6,7 @@
 #include "gzsize.h"
 
 
-long get_size(int fd) {
+long long get_size(int fd) {
         int ret;
         z_stream strm;
         unsigned char in[CHUNK];
